Replace viewer for 3d objects from 3dviewer.net to f3d.app
see f3d.app PR merged for parsing model url and extension: https://github.com/f3d-app/f3d/pull/1596
`https://f3d.app/web/#model=${modelUrl}`
